requirements.txt
# critical libs
Django==4.1.1
djangorestframework==3.15.1
django-simple-history==3.5.0
# django-background-tasks==1.2.5 # archived
# https://pypi.org/project/django4-background-tasks/
django4-background-tasks==1.2.9  # latest support django 4.1.1
# latest support django 4, migrate to django-tenants. Remove load_tenant_fixture afterward
gunicorn==23.0.0
django-tenant-schemas==1.12.0 # migrate to django-tenants and remove
django-tenants==3.7.0
celery==5.4.*
flower==2.0.*
redis==5.1.*
tenant-schemas-celery==3.0.*

# djangorestframework-jwt==1.11.0 # depricated
djangorestframework-simplejwt==5.3.1
django-otp==1.5.1
qrcode==7.4.2


psycopg==3.1.19
pandas==2.2.*
python-docx==1.1.2
six==1.*
requests==2.31.*
drf-yasg==1.21.7
django-cors-headers==4.3.1
python-dateutil==2.9.*
patchy==2.8.0
aiohttp==3.9.5

# tests
pytest==8.2.*
pytest-django==4.8.*
pytest-cov==5.0.*
pytest-xdist==3.6.1
pytest-sugar==1.0.0
Faker==25.3.*
time-machine==2.14.*
pytest-asyncio==0.23.7
pytest-factoryboy==2.7.0
### API
twilio
# refactor. as sendgrid now supports sms and emails in this package
sendgrid==6.11.*

# aws
boto3==1.34.97
botocore==1.34.97
watchtower==3.2.*
django-ses==4.0.*
django-storages==1.14.4
aws_secretsmanager_caching==1.1.*

sentry-sdk==2.18.0

### not yet updated
PyJWT==1.7.1

certifi==2019.11.28
chardet==3.0.4
django-allow-cidr==0.6.0
docusign-esign==3.25.1
docutils==0.15.2
entrypoints==0.3
future==0.18.2
google-auth==2.0.2
idna==2.8
importlib-metadata==4.13.0
jmespath==0.9.5
ordered-set==3.1.1
python-dotenv==0.11.0
python-http-client==3.2.7
xlrd==1.2.0
django-localflavor==4.0
django-extensions==3.2.3
django-filter==22.1


# linters, in sync with .pre-commit-config
black==24.4.2
black[d]  # daemon for local development
isort==5.13.2
pre-commit==3.5.0 # in sync with bitbucket pipeline
autoflake==2.3.1
flake8==7.0.0
django-oauth-toolkit==2.3.0

#swagger docs using drf_spectacular library
drf_spectacular_sidecar==2024.12.1
drf-spectacular==0.28.0
xhtml2pdf==0.2.17
